project/requirements.txt
Flask==3.0.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0

